fix valgrind memory leaks and a null pointer test
There's still a leak where arguments from .agrc are on the heap but
can't be freed during cleanup_options because we don't keep track of
where they end up in the final argv array.
Some refactoring would be needed to keep track of which options came
from .agrc and can be freed.
On the other hand, cleanup_options happens right before the program will
exit anyway, so the main reason for freeing everything is for
correctness. (also the bookkeeping data would take up additional memory)