boyer moore strnstr: fix pointer arithmetic unsigned overflow
If there's a match of a one or two character pattern at the beginning of
the string, pos ends up being -1 despite being an unsigned size_t. This
makes the undefined sanitizer in clang 5.0.0 complain:

  % echo 'foo bar' | ./ag fo
  ../src/util.c:187:22: runtime error: addition of unsigned offset to 0x557335584520 overflowed to 0x55733558451f
  foo bar

Fix this by making pos a signed ssize_t since it might become negative.
Cast s_len to an ssize_t when comparing to pos to fix a "comparison of
integers of different signs" warning