support for arbitrary regex patterns in the lang extension list
Running `ag --make` is basically useless out of the box because it only
looks for file extensions, so I added support for including arbitrary
regexes in language specifications.

When adding to langs[], if the first character in an "extension" is
a '^' then treat the rest of the string as a regex for the filename.

Unlike osimola's implementation which I tried previously, this method
doesn't change the definition of lang_spec_t, so merging new languages
from upstream is no longer a constant source of merge conflicts.

I simplified things by dropping combine_file_extensions and the crazy
static array things it was doing, so now parse_options only needs to
call one function to get the full regex.
Adding ag_dsprintf made the make_lang_regex code a lot easier to work
with and avoids unchecked sprintf calls which just hope that the buffer
is big enough.