Remove -Wno-implicit-fallthrough warning, suppress with comments in the code
It turns out that GCC can read comments as intentional fallthrough
markers (instead of just compiler attributes), so I can take out the
configure hack to check the GCC version.

Weirdness in options.c - the entire comment has to match a particular
"fall through" regex [1] so I had to remove the "so regex is
built" part of the comment to suppress the warning.

[1] https://gcc.gnu.org/onlinedocs/gcc/Warning-Options.html#index-Wimplicit-fallthrough