clang-format: check for 6.0 in configure, ignore usage text
clang-format doesn't like opening the usage text string literals on
column 0, but its suggested place in the middle of the page is silly.
Fortunatly simple comments can disable clang-format for a region of
code.

Also, look for clang-format-6.0 in the configure script for Debian boxes
which don't symlink 'clang-format' to the latest version like Arch does.