remove pcre1 support
I never use the pcre1 library, so it's advantageous to remove a layer of
abstraction and support only pcre2 going forward.

Possibly the biggest advantage here is that every call to ag_pcre_match
doesn't have to allocate and free the pcre2_match_data structure and
copy the ovector data; the same match data struct can be re-used for
multiple matches.

In a benchmark searching for a URL regex in the linux kernel tree with
filetypes enabled, there's a 20% end-to-end speedup from 1.73s down to 1.38s.

./ag --stats -c --workers=1 --cc --asm '[\w]+://[^/\s?#]+[^\s?#]+(?:\?[^\s#]*)?(?:#[^\s]*)?' /tmp/linux-src

While I'm here, I cleaned up checking for JIT (use it for all regexes
rather than just the main one).