configure: split up CFLAGS/CPPFLAGS and LDFLAGS/LIBS
This was inspired by seeing "rejected by the preprocessor but accepted
by the compiler" errors for headers in nonstandard paths, because when
doing preprocessor checks autoconf only included CPPFLAGS, but the -I
flags added by pkg-config checks went to CFLAGS. Similarly, -L and -l
options all end up in LIBS instead of being split up properly.

To handle this, write macros which split up the flags and append them to
the appropriate variable.
  * -I, -D, and -U options go to CPPFLAGS, everything else to CFLAGS
  * -l options go to LIBS and everything else to LDFLAGS
    * This might not work when certain ordering matters (e.g. with
      --as-needed or --whole-archive), but that doesn't apply to ag and
      is a PITA with automake anyway.
  * pull in AX_APPEND_FLAG from the autoconf archive which skips duplicates
  * use AC_CONFIG_MACRO_DIRS insetad of m4_include
  * bump AC_PREREQ to 2.64 to match AX_APPEND_FLAG
    * RHEL 6.10 has autoconf 2.65 and that's the oldest OS I care about