autotools: fix for old versions in RHEL 6
The aclocal from automake 1.11.1 (as seen in RHEL/CentOS 6.10) is wimpy
and can't figure out AC_CONFIG_MACRO_DIR(S) on its own, isntead we have
to explicitly pass "-I m4" to aclocal for it to find AX_PTHREAD and
every new macro that's mentioned after AX_PTHREAD.

Add this to autogen.sh and Makefile.am so that autogen and autoreconf
both work.

Change AC_CONFIG_MACRO_DIRS to AC_CONFIG_MACRO_DIR which is basically
the same but exists in older versions of autoconf.