drop zfile and fopencookie support
Opening compressed files as a stream with fopencookie is a neat idea,
but it's fundamentally incompatible with the way ag handles searching
and printing matches in multiple files.

When searching files, the full file data is available in a memory buffer
(either on the heap or through mmap). When we're done searching, the
print mutex is grabbed, and the file offsets are used to print all
matches for the whole file at once.

search_stream() handles printing totally differently. Matches are
printed as they're found while streaming (possibly along with context),
and the print mutex isn't touched. Theoretically we could buffer all the
matching lines and print them at once under the mutex, but that isn't
practical because of the more fundamental issue. Streams have no
definite endpoint; we could read an arbitrarily large amount of data
from a pipe. This doesn't work because the worker thread searching the
stream could hang indefinitely, and there would be no output at all
until the stream is closed.

Stream searching really only works if there's only one stream for ag to
search, usually stdin. The zfile code is messy, and a user wants to
stream a large compressed file, it'd be more practical to just pipe
a decompressor into ag.

I've had zfile disabled by default for almost two years, it's time to
just kill it.