search: fix /proc files that successfully seek to 0 length
Apparently lseek() to SEEK_END return 0 for some proc files, even though
there's actual data to be read. Ag was treating this as an empty file
that didn't need to be searched.

Fix by using the "read as much as we can and resize the buffer" strategy
for these files. Also change the initial size of the buffer to 1K rather
than 4K since most proc files are fairly small.