add -Z/--null-lines option to search files with null-separated "lines"
This is mainly for searching /proc/<pid>/environ since I got tired of
always piping through `tr '\0' '\n'`. For the sake of simplicity, this
option applies to all files searched, and actual '\n' characters are
ignored by ag (but could still be matched by the PCRE regex).

It's not very good when used in a directory containing normal text
files, because a matching null-separated "line" of an ASCII file is the
entire file. But still, this niche option has its place.