don't use asprintf when it's not needed
While profiling with valgrind/callgrind (which is extremely intrusive
and slow but provides data about libc functions), I found that
a significant fraction of search_dir was being spent in ag_asprintf
(really vasprintf and its internals).

Most of the ag_asprintf uses were to join paths, in the form of
ag_asprintf(&output, "%s/%s", path1, path2) which is straightforward
enough that it doesn't need a full printf call.

That particular printf format can be simplified down to a single malloc,
two strlen calls, and two memcpy calls, plus couple array element
accesses. I put that into a new join_paths() function which is used
wherever possible.

The end-to-end speedup is pretty small, but ag_asprintf is a lot less
hot now.